+++ src/constants/export.js
@@ -1,0 +1,16 @@
+export const EXPORT_JPG = 'jpeg'
+export const EXPORT_PNG = 'png'
+export const EXPORT_PDF = 'pdf'
+export const EXPORT_CSV = 'csv'
+export const EXPORT_OWL = 'owl'
+
+export const EXPORT_GRAPH_OPTIONS = [
+  EXPORT_JPG,
+  EXPORT_PNG,
+  EXPORT_PDF
+]
+
+export const EXPORT_DATA_OPTIONS = [
+  EXPORT_CSV,
+  EXPORT_OWL
+]