+++ src/utils/serialiseNodesEdges/getEdge.js
@@ -1,9 +1,24 @@
+/**
+ * Get edge info and related nodes
+ * @param  {Object}   params
+ * @param  {String}   params.from                    Subject node ID
+ * @param  {String}   params.predicate               Predicate node ID
+ * @param  {String}   params.to                      Object node ID
+ * @param  {Object}   params.classesFromApi          Nodes from initial OwlClasses
+ * @param  {Object}   params.objectPropertiesFromApi Edges from initial OwlObjectProperties
+ * @return {Object}   output
+ * @return {String}   output.edgeUniqueID            Edge unique ID
+ * @return {Object}   output.edgeConnection          Edge connection containing nodes IDs and labels
+ * @return {Object}   output.edge                    Edge object
+ * @return {Object}   output.fromObject              Subject node object
+ * @return {Object}   output.toObject                Object node object
+ */
 const getEdge = ({
+  classesFromApi,
   from,
+  objectPropertiesFromApi,
   predicate,
   to,
-  objectPropertiesFromApi,
-  classesFromApi
 }) => {
   const edgeUniqueId = `${predicate}___${from}___${to}`
   const edgeLabel = objectPropertiesFromApi[predicate].rdfsLabel